Cambiar el array de los productos y hacer un arraylist por cada producto

+++ src/Textil.java
@@ -1,7 +1,7 @@
 public class Textil extends Producto {
     protected String Composicio;
 
-    public Textil(String nomProducte, int preu, String codiBarras, String composicio) {
+    public Textil(String nomProducte, float preu, String codiBarras, String composicio) {
         super(nomProducte, preu, codiBarras);
         Composicio = composicio;
     }
